Fix critical security bugs, add pessimistic locking, decimal stock, and secure settings

+++ backend/products/models.py
@@ -16,7 +16,7 @@ class Product(models.Model):
     quality = models.CharField(max_length=20, choices=QUALITY_CHOICES)
     description = models.TextField(blank=True)
     image_urls = models.JSONField(default=list)  # آرایه‌ای از آدرس تصاویر
-    stock = models.IntegerField(default=0)
+    stock = models.DecimalField(max_digits=10, decimal_places=2, default=0)
     is_active = models.BooleanField(default=True)
     created_by = models.ForeignKey(User, on_delete=models.RESTRICT, related_name='products')
     created_at = models.DateTimeField(auto_now_add=True)
@@ -27,8 +27,8 @@ class Product(models.Model):
 
 class PriceHistory(models.Model):
     product = models.ForeignKey(Product, on_delete=models.RESTRICT, related_name='price_histories')
-    old_price = models.DecimalField(max_digits=15, decimal_places=2, null=True, blank=True)
-    new_price = models.DecimalField(max_digits=15, decimal_places=2)
+    old_price = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
+    new_price = models.DecimalField(max_digits=10, decimal_places=2)
     changed_by = models.ForeignKey(User, on_delete=models.RESTRICT, related_name='price_changes')
     changed_at = models.DateTimeField(auto_now_add=True)
 
@@ -44,8 +44,8 @@ class StockHistory(models.Model):
     )
     
     product = models.ForeignKey(Product, on_delete=models.RESTRICT, related_name='stock_histories')
-    old_stock = models.IntegerField(null=True, blank=True)
-    new_stock = models.IntegerField()
+    old_stock = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True) # اصلاح شد
+    new_stock = models.DecimalField(max_digits=10, decimal_places=2)                      # اصلاح شد
     reason = models.CharField(max_length=50, choices=REASON_CHOICES)
     changed_by = models.ForeignKey(User, on_delete=models.RESTRICT, related_name='stock_changes')
     changed_at = models.DateTimeField(auto_now_add=True)
